+++ wai-cryptocookie/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.1
+
+* Changed the type of `middleware` so that the previous *lookup*
+  function is not used anymore. Instead, a `CryptoCookie` is
+  provided directly to an `Application`-construction function.
+
 # Version 0.0.1
 
 * Initial version.